requirements.txt
async-lru==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pyaes==1.6.1
pycryptodome==3.18.0
pyrogram
pyromod==1.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
TgCrypto==1.2.5
urllib3==2.0.3
websockets==11.0.3
yt-dlp==2023.6.22
motor
aiohttp
aiofiles
pytz
ffmpeg
umongo==3.1.0
speedtest-cli
Flask==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
werkzeug==2.0.2
itsdangerous==2.0.1
m3u8
pycryptodomex